Fixed fatal bug,improvments to text editor. Compiles on mac.
Fixed bug with not comparing floating points in myjit by value and not bits. This lead to a bug when a sign mask was being intrepreted as 0.0 and cuases jit_negr to act up.
Compiles on macOS but im not providing binaries for it yet.
Can quit search/goto-line(new feature) with ctrl+'q'.